==> Projects/project_5-master/new_code/01_scraping_tweets_5.ipynb <==
## **Notebook Contents**
- Import Libraries
- Function to Grab Date, Tweet, & Username For the Queries
- Pull in Tweets
- Make Combined DataFrames
- Final Dataframe
- Adding Random Tweets


## **Import Libraries**

import GetOldTweets3 as got
import time
from datetime import datetime, date, timedelta
import pandas as pd
import numpy as np
import os
import tweepy as tw

## **Function to Grab Date, Tweet, & Username For the Queries**

# Source: https://github.com/Mottl/GetOldTweets3/issues/57
# Credit to Clairedevries

# Tweet scraper function
def DownloadTweets(SinceDate, UntilDate, Query):
    '''
    Downloads all tweets from a certain month in three sessions in order to avoid sending too many requests. 
    Date format = 'yyyy-mm-dd'. 
    '''
    since = datetime.strptime(SinceDate, '%Y-%m-%d')
    until= datetime.strptime(UntilDate, '%Y-%m-%d')
    tenth = since + timedelta(days = 10)
    twentieth = since + timedelta(days=20)
    
    first = got.manager.TweetCriteria().setQuerySearch(Query).setSince(since.strftime('%Y-%m-%d')).setUntil(tenth.strftime('%Y-%m-%d'))
    firstdownload = got.manager.TweetManager.getTweets(first)
    firstlist=[[tweet.date, tweet.text, tweet.username] for tweet in firstdownload]
    
    # The first 10 days of the month
    df_1 = pd.DataFrame.from_records(firstlist, columns = ["date", "tweet", "user"])

    # We set a 10 minute sleep timer to ensure a max iteration limit is not hit.
    time.sleep(600)
    
    second = got.manager.TweetCriteria().setQuerySearch(Query).setSince(tenth.strftime('%Y-%m-%d')).setUntil(twentieth.strftime('%Y-%m-%d'))
    seconddownload = got.manager.TweetManager.getTweets(second)
    secondlist=[[tweet.date, tweet.text, tweet.username] for tweet in seconddownload]
    
    # The second 10 days of the month
    df_2 = pd.DataFrame.from_records(secondlist, columns = ["date", "tweet", "user"])
    
    time.sleep(600)
    
    print ('Finishing the month')
    print('='*28)
    third = got.manager.TweetCriteria().setQuerySearch(Query).setSince(twentieth.strftime('%Y-%m-%d')).setUntil(until.strftime('%Y-%m-%d'))
    thirddownload = got.manager.TweetManager.getTweets(third)
    thirdlist=[[tweet.date, tweet.text, tweet.username] for tweet in thirddownload]
    
    # The rest of the month
    df_3 = pd.DataFrame.from_records(thirdlist, columns = ["date", "tweet", "user"])
    
    df=pd.concat([df_1,df_2,df_3])
  
    return df # The final dataframe of the month is returned

### **Pull In Tweets**

Queries:
- my power went out  
- i lost power  
- there is a power outage  
- I have no power  

# My Power Went Out
jan_fin_tweets_mpwo =  DownloadTweets('2020-01-01', '2020-01-31', 'my power went out')
time.sleep(100)
feb_fin_tweets_mpwo =  DownloadTweets('2020-02-01', '2020-02-29', 'my power went out')
time.sleep(100)
mar_fin_tweets_mpwo =  DownloadTweets('2020-03-01', '2020-03-31', 'my power went out')
time.sleep(100)
apr_fin_tweets_mpwo =  DownloadTweets('2020-04-01', '2020-04-30', 'my power went out')
time.sleep(100)
may_fin_tweets_mpwo =  DownloadTweets('2020-05-01', '2020-05-31', 'my power went out')
time.sleep(100)
jun_fin_tweets_mpwo =  DownloadTweets('2020-06-01', '2020-06-30', 'my power went out')
time.sleep(100)
jul_fin_tweets_mpwo =  DownloadTweets('2020-07-01', '2020-07-31', 'my power went out')
time.sleep(100)
aug_fin_tweets_mpwo =  DownloadTweets('2020-08-01', '2020-08-31', 'my power went out')
time.sleep(100)

# I lost power
jan_fin_tweets_ilp =  DownloadTweets('2020-01-01', '2020-01-31', 'i lost power')
time.sleep(100)
feb_fin_tweets_ilp =  DownloadTweets('2020-02-01', '2020-02-29', 'i lost power')
time.sleep(100)
mar_fin_tweets_ilp =  DownloadTweets('2020-03-01', '2020-03-31', 'i lost power')
time.sleep(100)
apr_fin_tweets_ilp =  DownloadTweets('2020-04-01', '2020-04-30', 'i lost power')
time.sleep(100)
may_fin_tweets_ilp =  DownloadTweets('2020-05-01', '2020-05-31', 'i lost power')
time.sleep(100)
jun_fin_tweets_ilp =  DownloadTweets('2020-06-01', '2020-06-30', 'i lost power')
time.sleep(100)
jul_fin_tweets_ilp =  DownloadTweets('2020-07-01', '2020-07-31', 'i lost power')
time.sleep(100)
aug_fin_tweets_ilp =  DownloadTweets('2020-08-01', '2020-08-31', 'i lost power')
tim.sleep(100)

# There is a power outage
jan_fin_tweets_tpo =  DownloadTweets('2020-01-01', '2020-01-31', 'there is a power outage')
time.sleep(100)
feb_fin_tweets_tpo =  DownloadTweets('2020-02-01', '2020-02-29', 'there is a power outage')
time.sleep(100)
mar_fin_tweets_tpo =  DownloadTweets('2020-03-01', '2020-03-31', 'there is a power outage')
time.sleep(100)
apr_fin_tweets_tpo =  DownloadTweets('2020-04-01', '2020-04-30', 'there is a power outage')
time.sleep(100)
may_fin_tweets_tpo =  DownloadTweets('2020-05-01', '2020-05-31', 'there is a power outage')
time.sleep(100)
jun_fin_tweets_tpo =  DownloadTweets('2020-06-01', '2020-06-30', 'there is a power outage')
time.sleep(100)
jul_fin_tweets_tpo =  DownloadTweets('2020-07-01', '2020-07-31', 'there is a power outage')
time.sleep(100)
aug_fin_tweets_tpo =  DownloadTweets('2020-08-01', '2020-08-31', 'there is a power outage')
time.sleep(100)

# I have no power
jan_fin_tweets_ihnp =  DownloadTweets('2020-01-01', '2020-01-31', 'I have no power')
time.sleep(100)
feb_fin_tweets_ihnp =  DownloadTweets('2020-02-01', '2020-02-29', 'I have no power')
time.sleep(100)
mar_fin_tweets_ihnp =  DownloadTweets('2020-03-01', '2020-03-31', 'I have no power')
time.sleep(100)
apr_fin_tweets_ihnp =  DownloadTweets('2020-04-01', '2020-04-30', 'I have no power')
time.sleep(100)
may_fin_tweets_ihnp =  DownloadTweets('2020-05-01', '2020-05-31', 'I have no power')
time.sleep(100)
jun_fin_tweets_ihnp =  DownloadTweets('2020-06-01', '2020-06-30', 'I have no power')
time.sleep(100)
jul_fin_tweets_ihnp =  DownloadTweets('2020-07-01', '2020-07-31', 'I have no power')
time.sleep(100)
aug_fin_tweets_ihnp =  DownloadTweets('2020-08-01', '2020-08-31', 'I have no power')

## **Make Combined DataFrames**

mpwo_final_df = jan_fin_tweets_mpwo.append([feb_fin_tweets_mpwo, mar_fin_tweets_mpwo, apr_fin_tweets_mpwo, 
                                            may_fin_tweets_mpwo, jun_fin_tweets_mpwo, jul_fin_tweets_mpwo, aug_fin_tweets_mpwo])

ilp_final_df = jan_fin_tweets_ilp.append([feb_fin_tweets_ilp, mar_fin_tweets_ilp, apr_fin_tweets_ilp, 
                                            may_fin_tweets_ilp, jun_fin_tweets_ilp, jul_fin_tweets_ilp, aug_fin_tweets_ilp])

tpo_final_df = jan_fin_tweets_tpo.append([feb_fin_tweets_tpo, mar_fin_tweets_tpo, apr_fin_tweets_tpo, 
                                            may_fin_tweets_tpo, jun_fin_tweets_tpo, jul_fin_tweets_tpo, aug_fin_tweets_tpo])

ihnp_final_df = jan_fin_tweets_ihnp.append([feb_fin_tweets_ihnp, mar_fin_tweets_ihnp, apr_fin_tweets_ihnp, 
                                            may_fin_tweets_ihnp, jun_fin_tweets_ihnp, jul_fin_tweets_ihnp, aug_fin_tweets_ihnp])

print(mpwo_final_df.shape)
print(ilp_final_df.shape)
print(tpo_final_df.shape)
print(ihnp_final_df.shape)

## **Final DataFrame**

ready_for_processing = mpwo_final_df.append([ilp_final_df,
                                             tpo_final_df, # This one will be nathans
                                             ihnp_final_df])

print(ready_for_processing.shape)

ready_for_processing.to_csv('./data/ready_for_processing.csv', index=False)

## **Adding Random Tweets**

# We went out shopping
jan_fin_tweets_wwos =  DownloadTweets('2020-03-01', '2020-03-31', 'we went out shopping')
time.sleep(100)
feb_fin_tweets_wwos =  DownloadTweets('2020-02-01', '2020-02-29', 'we went out shopping')
time.sleep(100)
mar_fin_tweets_wwos =  DownloadTweets('2020-03-01', '2020-03-31', 'we went out shopping')
time.sleep(100)
apr_fin_tweets_wwos =  DownloadTweets('2020-04-01', '2020-04-30', 'we went out shopping')
time.sleep(100)
may_fin_tweets_wwos =  DownloadTweets('2020-05-01', '2020-05-31', 'we went out shopping')
time.sleep(100)
jun_fin_tweets_wwos =  DownloadTweets('2020-06-01', '2020-06-30', 'we went out shopping')
time.sleep(100)
jul_fin_tweets_wwos =  DownloadTweets('2020-07-01', '2020-07-31', 'we went out shopping')
time.sleep(100)
aug_fin_tweets_wwos =  DownloadTweets('2020-08-01', '2020-08-31', 'we went out shopping')
time.sleep(100)

# The weather is nice
jan_tweets_twin =  DownloadTweets('2020-01-01', '2020-01-31', 'The weather is nice') 
time.sleep(100)
feb_tweets_twin =  DownloadTweets('2020-02-01', '2020-02-29', 'The weather is nice')   
time.sleep(100)
mar_tweets_twin =  DownloadTweets('2020-03-01', '2020-03-31', 'The weather is nice')  
time.sleep(100)
apr_tweets_twin =  DownloadTweets('2020-04-01', '2020-04-30', 'The weather is nice')  
time.sleep(100)
may_tweets_twin =  DownloadTweets('2020-05-01', '2020-05-31', 'The weather is nice')  
time.sleep(100)
jun_tweets_twin =  DownloadTweets('2020-06-01', '2020-06-30', 'The weather is nice')  
time.sleep(100)
jul_tweets_twin =  DownloadTweets('2020-07-01', '2020-07-31', 'The weather is nice')  
time.sleep(100)
aug_tweets_twin =  DownloadTweets('2020-08-01', '2020-08-31', 'The weather is nice') 

bad_words = jan_fin_tweets_wwos.append([feb_fin_tweets_wwos, mar_fin_tweets_wwos, apr_fin_tweets_wwos, may_fin_tweets_wwos,
                                       jun_fin_tweets_wwos, jul_fin_tweets_wwos, aug_fin_tweets_wwos])

# These will all be labeled with 0
bad_words.to_csv("./data/bad_words1.csv", index=False)

bad_words2 = jan_tweets_twin.append([feb_tweets_twin, mar_tweets_twin, apr_tweets_twin, may_tweets_twin,
                                       jun_tweets_twin, jul_tweets_twin, aug_tweets_twin])

# These will all be labeled with 0
bad_words2.to_csv("./data/bad_words2.csv", index=False)
