==> Projects/project_5-master/new_code/07_twitter_bot_evaluation.ipynb <==
## **Notebook Contents**
- Import Libraries
- Import Dataframes
- Uniform Labeling
- Model Evaluation Metrics


## **Import Libraries**

# Imports
import numpy as np
import pandas as pd

## **Import Dataframe**

# On new Tweets from pred_verify.csv
data = pd.read_csv('../data/pred_verify.csv')

data.head()

## **Uniform Labeling**

# Change random forest floats into 0's and 1's
ones_zeroes = []
for row in data['predict_rf']:
    if row > .50:
        ones_zeroes.append(1)
    else:
        ones_zeroes.append(0)

data['rf_10'] = ones_zeroes

## **Model Evaluation Metrics**

def acc(df, col):
    # Verified labels & labeled by script
    TP = []
    FP = []
    FN = []
    TN = []
    for index, row in df.iterrows():
        if row[col] and row['verify_labels'] == 1.0:
            TP.append(1)
        elif row[col] == 1.0 and row['verify_labels'] == 0.0:
            FP.append(1)
        elif row[col] == 0.0 and row['verify_labels'] == 1.0:
            FN.append(1)
        else:
            TN.append(1)
            
    TP = np.sum(TP)
    FP = np.sum(FP)
    FN = np.sum(FN)
    TN = np.sum(TN)
        
    # Accuracy:  # What percentage of observation did I correctly predict?
    print(f"The Accuracy score is: {round((TP + TN) / (TP + TN + FP + FN), 4) * 100}%")
    # Misclassification: # What percentage of observation did I incorrectly predict?
    print(f"The Missclassification rate is: {round((FP + FN) / (TP + TN + FP + FN), 3) * 100}%")
    # Sensitivity:  # Among Positives, how many did I get correctly? Same as recall
    print(f"The Sensitivity is: {round(TP / (TP + FN), 4) * 100}%")
    # Specificity:  # Among Negatives, how many did I get correctly?
    print(f"The Specificity is: {round(TN / (TN + FP), 4) * 100}%")
    # Precision: # Among Positives, how many did i Predict correct
    print(f"The Precision is {round(TP / (TP + FP), 4) * 100}%")

   
    return np.sum(TP), np.sum(FP), np.sum(FN), np.sum(TN)

# Logistic Regression
acc(data, 'predict_lr')

# Multinominal Naive Bayes
acc(data, 'predict_nb')

# LSTM RNN
acc(data, 'predict_rnn')

# Random Forest
acc(data, 'rf_10')

# Multi-Layered Model
acc(data, 'labels')
